% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_specs_list.r
\name{aggregate_specs_list}
\alias{aggregate_specs_list}
\title{aggregate_specs_list}
\usage{
aggregate_specs_list(sl, byFeature = FALSE, fd = NULL, fd_id = 1)
}
\arguments{
\item{sl}{specs_list. A named list of outputs from phy_or_env_spec. See examples.}

\item{byFeature}{bool. If true, each feature will occupy only one row, with multiple
columns to represent the different variables in specs_list (default: FALSE)}

\item{fd}{data.frame. Optional feature data - a data.frame object with one row per
feature, including some column with feature IDs that includes feature IDs in sl as
rownames (default:NULL)}

\item{fd_id}{integer or string. If integer, specifies the column index of fd that
contains feature ids. If character, specifies the column name (default: 1).}
}
\value{
a data.frame object.
}
\description{
Aggregates a list of outputs from specificity::phy_or_env_spec() into a single
data.frame object. Can also include feature data (e.g. species taxonomy) into
that output. Output can also be byFeature, with one row per feature, and multiple
columns for different variables.
}
\examples{
# attach(antarctica)
# # to see details of the antarctica data set, see ?antarctica
# # aggregate data long format:
# ant_long <- aggregate_specs_list(antarctica_specs_list, byFeature=FALSE,
#   fd=antarctica_taxonomy, fd_id=1)
# head(ant_long)
# ant_wide <- aggregate_specs_list(antarctica_specs_list, byFeature=TRUE,
#   fd=antarctica_taxonomy, fd_id=1)
# head(ant_wide)
# # aggregation can also be done without feature data (fd):
# ant_long <- aggregate_specs_list(antarctica_specs_list, byFeature=FALSE)
# head(ant_long)
# ant_wide <- aggregate_specs_list(antarctica_specs_list, byFeature=TRUE)
# head(ant_wide)


}
\author{
John L. Darcy
}
