% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_specs_shiny.r
\name{plot_specs_shiny}
\alias{plot_specs_shiny}
\title{plot_specs_shiny}
\usage{
plot_specs_shiny(sl, fd = NULL, fd_id_col = 1)
}
\arguments{
\item{sl}{specs_list. A named list of outputs from phy_or_env_spec. See examples.}

\item{fd}{data.frame. Optional feature data - a data.frame object with one row per
feature, including some column with feature IDs that includes feature IDs in sl as
rownames. If NULL, no feature data will be used (default:NULL).}

\item{fd_id_col}{integer or string. Either the column number of fd containing feature
IDs (see above), or the column name of that column (default: 1).}
}
\value{
Returns output of shiny::shinyApp(). If run locally, this will manifest
as a browser window opening with an interactive visualization.
}
\description{
Runs an interactive shiny visualization. Data can be re-downloaded from that
visualization, in either "wide" or "long" format. See ?aggregate_specs_list
for explanation of those formats.
}
\examples{
# attach(antarctica)
# plot_specs_shiny(antarctica_specs_list, antarctica_taxonomy, 1)


}
\author{
John L. Darcy
}
