% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{specShinyFuns}
\alias{specShinyFuns}
\title{Portable source code for specificity.shiny functions}
\format{
A list containing 2 objects:
\describe{
\item{plot_specs_shiny:}{
Source code for plot_specs_shiny(). Can be written to file using writeLines().
}
\item{aggregate_specs_list:}{
Source code for aggregate_specs_list(). Can be written to file using writeLines().
}
}
}
\usage{
specShinyFuns
}
\description{
This source code is included so that functions can be written out as .r files,
so that a server running the shiny visualization does not need to have the
specificity.shiny R package installed.
}
\keyword{datasets}
